#1 修复消息过滤tag失败的问题，原因是 long与 int 的问题， 通过long类型查询set  ，如果set中是存放的int类型数据，则查询不到

+++ rocketmq-store/src/main/java/com/alibaba/rocketmq/store/DefaultMessageFilter.java
@@ -23,7 +23,7 @@ public class DefaultMessageFilter implements MessageFilter {
             return true;
         }
 
-        return subscriptionData.getCodeSet().contains(tagsCode);
+        return subscriptionData.getCodeSet().contains((int) tagsCode);
     }
 
 }
